socket.cpp: Added a cast to signed int to the returning of data.size()
This prevent a warning message from g++(1)

Acked-by: Ricardo T Oliveira <rto07@c3sl.ufpr.br>
Acked-by: Vinicius Ruoso <vkr07@c3sl.ufpr.br>

+++ src/socket.cpp
@@ -38,7 +38,7 @@ SocketClient::SocketClient(int port, string h) {
 void SocketClient::SendMessage(string data) {
     char buffer[BUFSIZ];
 
-    if(write(sock_send, data.c_str(), data.size()) != data.size()) {
+    if(write(sock_send, data.c_str(), data.size()) != (int)data.size()) {
         ERROR("%s\n",strerror(errno));
         exit(1);
     }
